+++ src/operations_generales.h
@@ -1,11 +1,13 @@
 #ifndef OPERATIONS_GENERALES_H
 #define OPERATIONS_GENERALES_H
 
+#include <LibRobus.h>
+
 //Prend un tableau d'int et sa longueur, et classe les membres du tableau en ordre croissant
-void classerTableauCroissant(int tableau[], int NB_ELEMENTS);
+void classerTableauCroissant(uint16_t tableau[], int NB_ELEMENTS);
 
 //Prend un tableau d'int et sa longueur (impaire), et renvoie la mediane parmi ses membres
-float calculerMediane(int tableau[], int NB_ELEMENTS);
+uint16_t calculerMediane(uint16_t tableau[], int NB_ELEMENTS);
